requirements.txt
appdirs==1.4.3
asn1crypto==1.2.0
astroid==2.3.2
attrs==19.1.0
Automat==0.8.0
bottle==0.12.17
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
configparser==4.0.2
constantly==15.1.0
coverage==4.5.4
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
Django==2.2.6
django-appconf==1.0.3
django-autoslug==1.9.6
django-compressor==2.3
django-debug-toolbar==2.0
django-haystack==2.8.1
django-ipware==2.1.0
django-mdeditor==0.1.16
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-uuslug==1.1.9
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
et-xmlfile==1.0.1
Flask==1.1.1
freeze==1.0.10
get==2019.4.13
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipaddress==1.0.22
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.39
Jinja2==2.10.3
jsonpickle==1.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lxml==4.4.1
markdown2==2.3.8
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mysqlclient==1.4.4
numpy==1.17.3
olefile==0.46
openpyxl==3.0.1
packaging==19.2
pandas==0.25.2
parsel==1.5.2
pbr==5.4.3
Pillow==6.2.1
post==2019.4.13
Protego==0.1.15
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.4.3
pymongo==3.9.0
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-logstash==0.4.6
python-memcached==1.59
python-slugify==4.0.0
pytz==2019.3
query-string==2019.4.13
queuelib==1.5.0
raven==6.10.0
rcssmin==1.0.6
request==2019.4.13
requests==2.22.0
rjsmin==1.1.0
Scrapy==1.8.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.31.0
text-unidecode==1.3
Twisted==19.10.0
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.6
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
w3lib==1.21.0
webencodings==0.5.1
Werkzeug==0.16.0
WeRoBot==1.10.1
Whoosh==2.7.4
wrapt==1.11.2
xlrd==1.2.0
xmltodict==0.12.0
zope.interface==4.7.1
